change AnalysisTask.percentiles to int array
All existing RegionalAnalysis instances in Mongo have a
one-element double array holding an integer value.
Numbers sent over the API are always whole, so should
deserialize fine in old workers expecting doubles and new
workers expecting ints.

+++ src/main/java/com/conveyal/r5/analyst/cluster/AnalysisTask.java
@@ -76,7 +76,7 @@ public abstract class AnalysisTask extends ProfileRequest {
      * Which percentiles of travel time to calculate.
      * These should probably just be integers, but there are already a lot of them in Mongo as floats.
      */
-    public double[] percentiles;
+    public int[] percentiles;
 
     /**
      * The travel time cutoffs in minutes for regional accessibility analysis.
